feat: move Tip struct to types::parameters::tip
docs: rename docs for actions
