+++ src/types/chat/action.rs
@@ -15,9 +15,9 @@ pub enum Action {
     RecordVideo,
     /// About to send a video.
     UploadVideo,
-    /// About to send a generated voice-audio.
+    /// About to send a generated voice message.
     RecordVoice,
-    /// About to send a voice-audio.
+    /// About to send a voice message.
     UploadVoice,
     /// About to send a document.
     UploadDocument,
